requirements.txt
streamlit==1.32.0
openai==1.74.0
supabase==1.2.0
pandas==2.2.3
plotly==5.19.0
python-dotenv==1.1.0
requests==2.32.3
numpy==1.26.4
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0 